requirements.txt
flask==3.0.3
flask-socketio==5.3.6
eventlet==0.36.1
pyserial==3.5